+++ src/Properties/AssemblyInfo.cs
@@ -30,5 +30,5 @@ using System.Runtime.InteropServices;
 //
 // You can specify all the values or you can default the Revision and Build Numbers 
 // by using the '*' as shown below:
-[assembly: AssemblyVersion("0.9.21161.1816")]
-[assembly: AssemblyFileVersion("0.9.21161.1816")]
+[assembly: AssemblyVersion("0.9.21161.1831")]
+[assembly: AssemblyFileVersion("0.9.21161.1831")]