+++ src/Resources/Doc/changelog.md
@@ -3,7 +3,12 @@
 ## Version 0.9.x.x
 > Porting v0.8 over to use OptionObject2015
 
-#### 0.9.21161.18160.9.21161.1749 (2021-06-10)
+#### 0.9.21161.1831 (2021-06-10)
+* `ADDED` MyAvatoolWebService.GetVersion()
+* `ADDED` MyAvatoolWebService.RunScript()
+* `ADDED` MyAvatoolWebService.MethodName()
+
+#### 0.9.21161.1816 (2021-06-10)
 * `INFO` Added the NTST.ScriptLinkService.Objects project to the solution
 * `MODIFIED` MAWS Manual updates